+++ crates/rust-client/src/lib.rs
@@ -203,7 +203,7 @@ pub mod testing {
 use alloc::sync::Arc;
 
 use miden_lib::utils::ScriptBuilder;
-use miden_objects::crypto::rand::FeltRng;
+use miden_objects::{block::BlockNumber, crypto::rand::FeltRng};
 use miden_tx::{LocalTransactionProver, auth::TransactionAuthenticator};
 use rand::RngCore;
 use rpc::NodeRpcClient;
@@ -273,7 +273,7 @@ where
     /// # Errors
     ///
     /// Returns an error if the client couldn't be instantiated.
-    pub fn new(
+    pub async fn new(
         rpc_api: Arc<dyn NodeRpcClient + Send>,
         rng: Box<dyn FeltRng>,
         store: Arc<dyn Store>,
@@ -281,11 +281,16 @@ where
         exec_options: ExecutionOptions,
         tx_graceful_blocks: Option<u32>,
         max_block_number_delta: Option<u32>,
-    ) -> Self {
+    ) -> Result<Self, ClientError> {
         let authenticator = Some(authenticator);
         let tx_prover = Arc::new(LocalTransactionProver::default());
 
-        Self {
+        if let Some((genesis, _)) = store.get_block_header_by_num(BlockNumber::GENESIS).await? {
+            // Set the genesis commitment in the RPC API client for future requests.
+            rpc_api.set_genesis_commitment(genesis.commitment()).await?;
+        }
+
+        Ok(Self {
             store,
             rng: ClientRng::new(rng),
             rpc_api,
@@ -294,7 +299,7 @@ where
             exec_options,
             tx_graceful_blocks,
             max_block_number_delta,
-        }
+        })
     }
 
     /// Returns true if the client is in debug mode.