+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 
 ### Changes
 
-- [BREAKING] Incremented MSRV to 1.88.
 * [BREAKING] Incremented MSRV to 1.88.
 * [BREAKING] Updated `toBech32` AccountID method: it now expects a parameter to specify the NetworkID [(#1043)](https://github.com/0xMiden/miden-client/pull/1043).
 * Introduced enums instead of booleans for public APIs (#1042).
@@ -15,7 +14,7 @@
 
 * Added Ability to Convert Word to U64 Array and Felt Array in Web Client (#1041).
 * Added `TokenSymbol` Type to Web Client (#1046).
-* Added genesis commitment header to `TonicRpcClient` requests (#1045).
+* [BREAKING] Added genesis commitment header to `TonicRpcClient` requests (#1045).
 
 ## 0.10.1 (2025-07-26)